feat: two-server infrastructure with path-based routing /screenrecorder
- Configure two staging environments on separate servers:
  test (shepaland-cloud) at services-test.shepaland.ru/screenrecorder
  prod (shepaland-videocalls-test-srv) at services.shepaland.ru/screenrecorder
- Add Traefik StripPrefix middleware and per-env ingress/configmaps
- Add Vite base path, React Router basename, dynamic Axios BASE_URL
- Add nginx proxy routes for control-plane, ingest-gateway, minio
- Add MinIO k8s manifests (deployment, service, pvc, init-job)
- Add NetworkPolicies (default-deny + per-service)
- Fix security: timing-safe API key comparison, tenant_id isolation
- Increase liveness probe delays for memory-constrained nodes
- Add deploy scripts and Windows Agent testing guide

+++ ingest-gateway/src/main/java/com/prg/ingest/dto/request/PresignRequest.java
@@ -33,7 +33,7 @@ public class PresignRequest {
 
     @NotNull(message = "duration_ms is required")
     @Min(value = 1, message = "duration_ms must be > 0")
-    private Long durationMs;
+    private Integer durationMs;
 
     @NotBlank(message = "checksum_sha256 is required")
     private String checksumSha256;
